# Downloads/pulse 앱/pulse/event-collector/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Pulse Event Collector API
  version: 1.1.0
  description: First-party event collection, EQS metrics, funnel/cohort/path analytics.
servers:
  - url: http://localhost:8081
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
  /v1/events:
    post:
      summary: Ingest a first-party event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventPayload'
      responses:
        '202':
          description: Event accepted
        '400':
          description: Validation error
  /v1/metrics/pages:
    get:
      summary: Get page-level EQS metrics
      parameters:
        - in: query
          name: from
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: site_id
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
  /v1/funnels:
    get:
      summary: List funnel definitions
      parameters:
        - in: query
          name: site_id
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
    post:
      summary: Create or update funnel definition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FunnelDefinition'
      responses:
        '200':
          description: Saved
  /v1/funnels/{funnel_id}/report:
    post:
      summary: Query funnel report
      parameters:
        - in: path
          name: funnel_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DateRangeRequest'
      responses:
        '200':
          description: Funnel report
  /v1/cohorts:
    get:
      summary: List cohort definitions
      parameters:
        - in: query
          name: site_id
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
    post:
      summary: Create or update cohort definition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CohortDefinition'
      responses:
        '200':
          description: Saved
  /v1/cohorts/{cohort_id}/refresh:
    post:
      summary: Refresh one cohort snapshot
      parameters:
        - in: path
          name: cohort_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DateRangeRequest'
      responses:
        '200':
          description: Refreshed
  /v1/cohorts/refresh:
    post:
      summary: Refresh all cohort snapshots
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DateRangeRequest'
      responses:
        '200':
          description: Refreshed
  /v1/paths/query:
    post:
      summary: Query top conversion/revenue paths
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PathQuery'
      responses:
        '200':
          description: Top paths
components:
  schemas:
    EventPayload:
      type: object
      required:
        - site_id
        - session_id
        - event_name
        - ts
        - properties
        - consent_state
      properties:
        site_id:
          type: string
        session_id:
          type: string
        event_name:
          type: string
          description: Revenue standard events are purchase/subscription_start/donation.
        ts:
          oneOf:
            - type: string
              format: date-time
            - type: number
        properties:
          type: object
          additionalProperties: true
          description: For revenue events include amount, currency, product?, payment_provider?.
        consent_state:
          type: string
          enum: [granted, denied, unknown]
        idempotency_key:
          type: string
    FunnelDefinition:
      type: object
      required:
        - site_id
        - name
        - steps
        - conversion_window_minutes
      properties:
        funnel_id:
          type: string
        site_id:
          type: string
        name:
          type: string
        steps:
          type: array
          items:
            type: string
          description: event_name or page_path(/...)
        conversion_window_minutes:
          type: integer
    CohortDefinition:
      type: object
      required:
        - site_id
        - name
        - dsl
      properties:
        cohort_id:
          type: string
        site_id:
          type: string
        name:
          type: string
        dsl:
          type: object
          required:
            - all
          properties:
            all:
              type: array
              items:
                oneOf:
                  - type: object
                    required: [type, value]
                    properties:
                      type:
                        type: string
                        enum: [visit_count_gte]
                      value:
                        type: integer
                  - type: object
                    required: [type, event_name]
                    properties:
                      type:
                        type: string
                        enum: [has_event]
                      event_name:
                        type: string
                  - type: object
                    required: [type, value]
                    properties:
                      type:
                        type: string
                        enum: [eqs_gte]
                      value:
                        type: number
                  - type: object
                    required: [type, values]
                    properties:
                      type:
                        type: string
                        enum: [utm_source_in]
                      values:
                        type: array
                        items:
                          type: string
    PathQuery:
      type: object
      required:
        - site_id
        - from
        - to
        - start_event
        - end_event
      properties:
        site_id:
          type: string
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        start_event:
          type: string
        end_event:
          type: string
        top_n:
          type: integer
          default: 20
        max_path_length:
          type: integer
          default: 8
        sample_rate:
          type: number
          default: 1
        event_fetch_limit:
          type: integer
          default: 300000
    DateRangeRequest:
      type: object
      required:
        - from
        - to
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        site_id:
          type: string